Extract FidoState and frozen dataclasses into fido.appstate leaf module
All frozen dataclasses that compose FidoState (WorkerActivity, WorkerCrash,
WebhookActivity, ProviderSnapshot, ThreadSnapshot, RepoState, FidoState) and
create_fido_atomic() are now in fido.appstate — a true leaf module with no
back-reference to registry.py or the provider layer. Provider and session
modules (session_agent.py, claude.py, codex.py, copilotcli.py,
provider_factory.py) now import FidoState directly at runtime, eliminating the
TYPE_CHECKING guards. registry.py re-uses the types via a direct import;
private sentinels _ZERO_CRASH and _zero_activity stay in registry.py.

Helped-by: Robert Hencke <178348+rhencke@users.noreply.github.com>
