+++ src/fido/codex.py
@@ -10,9 +10,10 @@ from collections.abc import Callable, Iterable
 from dataclasses import dataclass
 from datetime import UTC, datetime
 from pathlib import Path
-from typing import IO, TYPE_CHECKING, Any, NoReturn, Protocol
+from typing import IO, Any, NoReturn, Protocol
 
 from fido import provider
+from fido.appstate import FidoState
 from fido.atomic import AtomicUpdater
 from fido.idle_timeout import IdleDeadline
 from fido.provider import (
@@ -31,9 +32,6 @@ from fido.provider import (
     model_name,
 )
 from fido.session_agent import SessionBackedAgent
-
-if TYPE_CHECKING:
-    from fido.registry import FidoState
 
 log = logging.getLogger(__name__)
 
@@ -1143,7 +1141,7 @@ class CodexClient(SessionBackedAgent, ProviderAgent):
         work_dir: Path | str | None = None,
         repo_name: str | None = None,
         session: PromptSession | None = None,
-        state_updater: "AtomicUpdater[FidoState] | None" = None,
+        state_updater: AtomicUpdater[FidoState] | None = None,
     ) -> None:
         self._runner = runner
         self._session_factory = (
