+++ src/fido/session_agent.py
@@ -5,8 +5,8 @@ import logging
 import threading
 from collections.abc import Callable
 from pathlib import Path
-from typing import TYPE_CHECKING
 
+from fido.appstate import FidoState
 from fido.atomic import AtomicUpdater
 from fido.provider import (
     READ_ONLY_ALLOWED_TOOLS,
@@ -14,9 +14,6 @@ from fido.provider import (
     ProviderModel,
     TurnSessionMode,
 )
-
-if TYPE_CHECKING:
-    from fido.registry import FidoState
 
 log = logging.getLogger(__name__)
 
@@ -35,7 +32,7 @@ class SessionBackedAgent:
         work_dir: Path | str | None,
         repo_name: str | None,
         session: PromptSession | None,
-        state_updater: "AtomicUpdater[FidoState] | None" = None,
+        state_updater: AtomicUpdater[FidoState] | None = None,
     ) -> None:
         self._session_fn = session_fn
         self._session_system_file = session_system_file
@@ -43,10 +40,10 @@ class SessionBackedAgent:
         self._repo_name = repo_name
         self._session_lock = threading.Lock()
         self._session: PromptSession | None = session
-        self._state_updater: "AtomicUpdater[FidoState] | None" = state_updater
+        self._state_updater: AtomicUpdater[FidoState] | None = state_updater
 
     @property
-    def state_updater(self) -> "AtomicUpdater[FidoState] | None":
+    def state_updater(self) -> AtomicUpdater[FidoState] | None:
         """Return the injected :class:`~fido.atomic.AtomicUpdater`, if any."""
         return self._state_updater
